requirements.txt
selenium==4.18.1
webdriver-manager==4.0.1
requests==2.31.0
pillow==10.2.0
pystray==0.19.5
pywin32==306
pyinstaller==6.5.0 